* commit 'e82b07e9bf25de558a1d3cb2d66f954a750c502e':
  Unified RenderScript logic between different calculators, and fixed some logic issues in MSSIM calculator

+++ tests/tests/uirendering/src/android/uirendering/cts/differencecalculators/ExactComparer.rs
@@ -9,7 +9,7 @@ rs_allocation given;
 
 // This method does a simple comparison of all the values in the given and ideal allocations.
 // If any of the pixels are off, then the test will fail.
-void exactCompare(const int32_t *v_in, int32_t *v_out){
+void exactCompare(const int32_t *v_in, float *v_out){
     int y = v_in[0];
     v_out[0] = 0;
